requirements.txt
django>=4.2,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
redis>=5.0,<6.0
google-generativeai>=0.5,<1.0
daphne>=4.0,<5.0